pdf_translator_base/pdf_translator_base/requirements.txt
Flask==3.1.2
google-genai>=1.40.0
pypdf==6.1.1
reportlab==4.4.4
python-dotenv==1.1.1
gunicorn==23.0.0
pytest==8.4.2
